Add files via upload
This is some of my work with python. This was a group project in collaboration with 2 students in my class whose names are listed in the author comment of the python files This is a command-line program that works with SQLite to create a simple flight database and read in from a CSV file. To run, save the files into a separate folder and run the program from the CLI.py file and the flightsinput.txt is the example of commands that can be run on the command line.